skycli/pyproject.toml
[project]
name = "skycli"
version = "0.1.0"
description = "Skyvault CLI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "grpcio>=1.71.0",
    "protobuf>=6.31.0",
    "fire>=0.7.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.7",
]
